requirements.txt
python-dotenv
python-telegram-bot==20.7
ADMIN_ID=770401066